Moved IAppState to /helpers/types.ts
