+++ src/engines/index.ts
@@ -1,5 +1,5 @@
 import { IRpcEngine } from "../helpers/types";
-import { IAppState } from "../App";
+import { IAppState } from "../helpers/types";
 import ethereum from "./ethereum";
 
 class RpcEngine implements IRpcEngine {
